ztp: tests update for policyGenerator
    - reorganized test files splitting constants, getters, assertions & helpers into their own files
    - updated test templates per changes made to example templates
    - modify policyGen test assertions as per changes made to template
    - preserve individual files for generated CRs that can be applied directly to the cluster
    - nbde siteconfig test fixes
    - alternative implementation for os.Executable for tests
    - ztp: gofmt fixes
    - removed the test flag approach
    - added a failsafe within fileHandler.go:ReadResourceFile
    - review fixes
    - review change to avoid loop for array of size 1
