+++ ztp/ztp-policy-generator/kustomize/plugin/policyGenerator/v1/policygenerator/testData/TestMachineConfigPool/sourcePolicies/TestMachineConfigPool.yaml
