+++ ztp/ztp-policy-generator/kustomize/plugin/policyGenerator/v1/policygenerator/testData/TestNamespace/sourcePolicies/TestNamespace.yaml
