+++ ztp/ztp-policy-generator/kustomize/plugin/policyGenerator/v1/policygenerator/testData/TestOperatorGroup/sourcePolicies/TestOperatorGroup.yaml
