+++ ztp/ztp-policy-generator/kustomize/plugin/policyGenerator/v1/policygenerator/testData/TestOperatorGroup/templates/TestOperatorGroup.yaml
@@ -1,0 +1,16 @@
+---
+apiVersion: policyGenerator/v1
+kind: PolicyGenTemplate
+metadata:
+  # The name will be used to generate the placementBinding and placementRule names as {name}-placementBinding and {name}-placementRule
+  name: "common-policies"
+  namespace: "policy-templates"
+spec:
+  bindingRules:
+    common: "true"
+  sourceFiles:
+    # Create operators policies that will be installed in all clusters
+    - fileName: TestOperatorGroup.yaml
+      policyName: "test-sub-oper-policy"
+    - fileName: TestOperatorGroup.yaml
+      policyName: ""
