+++ ztp/ztp-policy-generator/kustomize/plugin/policyGenerator/v1/policygenerator/testData/TestPtpConfig/sourcePolicies/TestPtpConfig.yaml
