+++ ztp/ztp-policy-generator/kustomize/plugin/policyGenerator/v1/policygenerator/testData/TestPtpConfig/templates/TestPtpConfig.yaml
@@ -1,0 +1,36 @@
+---
+apiVersion: policyGenerator/v1
+kind: PolicyGenTemplate
+metadata:
+  # The name will be used to generate the placementBinding and placementRule names as {name}-placementBinding and {name}-placementRule
+  name: "group-du-policies"
+  namespace: "policy-template"
+  labels:
+    common: false
+    groupName: "group-du"
+    siteName: "N/A"
+    mcp: "worker-du"
+spec:
+  bindingRules:
+    common: "true"
+  sourceFiles:
+    - fileName: TestPtpConfig.yaml
+      policyName: "ptp-config-policy"
+      metadata:
+        name: "du-ptp-slave"
+      spec:
+        profile:
+          - name: "slave"
+            interface: "ens5f0"
+            ptp4lOpts: "-2 -s --summary_interval -4"
+            phc2sysOpts: "-a -r -n 24"
+    - fileName: TestPtpConfig.yaml
+      policyName: ""
+      metadata:
+        name: "du-ptp-slave"
+      spec:
+        profile:
+          - name: "slave"
+            interface: "ens5f0"
+            ptp4lOpts: "-2 -s --summary_interval -4"
+            phc2sysOpts: "-a -r -n 24"
