+++ ztp/ztp-policy-generator/kustomize/plugin/policyGenerator/v1/policygenerator/testData/TestSubscription/sourcePolicies/TestSubscription.yaml
