+++ ztp/ztp-policy-generator/kustomize/plugin/policyGenerator/v1/policygenerator/testData/TestSubscription/templates/TestSubscription.yaml
@@ -5,12 +5,13 @@ metadata:
   # The name will be used to generate the placementBinding and placementRule names as {name}-placementBinding and {name}-placementRule
   name: "common-policies"
   namespace: "policy-templates"
-  labels:
-    common: true
-    groupName: "N/A"
-    siteName: "N/A"
-    mcp: "N/A"
-sourceFiles:
-  # Create operators policies that will be installed in all clusters
-  - fileName: TestNamespace
-    policyName: "test-sub-ns-policy"
+spec:
+  bindingRules:
+    common: "true"
+  sourceFiles:
+    # Create operators policies that will be installed in all clusters
+    - fileName: TestSubscription.yaml
+      policyName: "test-sub-policy"
+    - fileName: TestSubscription.yaml
+      policyName: ""
+
