--- ztp/ztp-policy-generator/kustomize/plugin/policyGenerator/v1/policygenerator/testFiles/TestOperatorGroup/templates/TestOperatorGroup.yaml
@@ -1,16 +1,0 @@
----
-apiVersion: policyGenerator/v1
-kind: PolicyGenTemplate
-metadata:
-  # The name will be used to generate the placementBinding and placementRule names as {name}-placementBinding and {name}-placementRule
-  name: "common-policies"
-  namespace: "policy-templates"
-  labels:
-    common: true
-    groupName: "N/A"
-    siteName: "N/A"
-    mcp: "N/A"
-sourceFiles:
-  # Create operators policies that will be installed in all clusters
-  - fileName: TestOperatorGroup
-    policyName: "test-sub-oper-policy"
