--- ztp/ztp-policy-generator/kustomize/plugin/policyGenerator/v1/policygenerator/testFiles/TestSriovNetwork/templates/TestSriovNetwork.yaml
@@ -1,20 +1,0 @@
----
-apiVersion: policyGenerator/v1
-kind: PolicyGenTemplate
-metadata:
-  # The name will be used to generate the placementBinding and placementRule names as {name}-placementBinding and {name}-placementRule
-  name: "site-du-sno-1-policies"
-  namespace: "policy-template"
-  labels:
-    common: false
-    groupName: "N/A"
-    siteName: "site-du-sno-1"
-    mcp: master
-sourceFiles:
-  - fileName: TestSriovNetwork
-    policyName: "sriov-nw-fh-policy"
-    metadata:
-      name: "sriov-nw-du-fh"
-    spec:
-      resourceName: du_fh
-      vlan: 140
