+++ ztp/ztp-policy-generator/kustomize/plugin/policyGenerator/v1/policygenerator/testSource/TestConstants.go
@@ -1,0 +1,18 @@
+package testSource
+
+const TEST_DIRECTORY = "testData"
+const OUT_DIRECTORY = "out"
+const TEMPLATE_DIRECTORY = "templates"
+const SOURCE_POLICIES_DIRECTORY = "sourcePolicies"
+const COMMON_DIRECTORY = "common"
+const GROUP_DIRECTORY = "groups"
+const SITE_DIRECTORY = "sites"
+const CUSTOM_RESOURCE_DIRECTORY = "customResource"
+const POLICIES = "-policies"
+const PLACEMENT_RULES = "placementrules.yaml"
+const PLACEMENT_BINDING = "placementbinding.yaml"
+const SPEC = "spec"
+const POLICY_TEMPLATES = "policy-templates"
+const OBJECT_TEMPLATES = "object-templates"
+const OBJECT_DEFINITION = "objectDefinition"
+const NOT_APPLICABLE = "N/A"
